Added inventory files and modified order process and database

+++ backend/models/orders.py
@@ -16,11 +16,9 @@ class Order(Base):
     
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
 
-    # Relationships
     items = relationship("OrderItem", back_populates="order", cascade="all, delete-orphan")
     status_history = relationship("OrderStatusHistory", back_populates="order", cascade="all, delete-orphan")
     delivery_info = relationship("OrderDelivery", back_populates="order", uselist=False, cascade="all, delete-orphan")
-
 
 class OrderStatusHistory(Base):
     __tablename__ = "order_status_history"
@@ -28,11 +26,10 @@ class OrderStatusHistory(Base):
     id = Column(Integer, primary_key=True, index=True)
     order_id = Column(Integer, ForeignKey("orders.id"), nullable=False)
     
-    status = Column(String, nullable=False) # Pending, Packed, Out for Delivery, Delivered, Abort
+    status = Column(String, nullable=False)
     changed_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
 
     order = relationship("Order", back_populates="status_history")
-
 
 class OrderDelivery(Base):
     __tablename__ = "order_deliveries"
@@ -40,12 +37,11 @@ class OrderDelivery(Base):
     id = Column(Integer, primary_key=True, index=True)
     order_id = Column(Integer, ForeignKey("orders.id"), unique=True, nullable=False)
     
-    driver_name = Column(String) # Could be a ForeignKey to an Employee table later
+    driver_name = Column(String) 
     customer_name = Column(String, nullable=False)
     delivery_address = Column(String, nullable=False)
     
     order = relationship("Order", back_populates="delivery_info")
-
 
 class OrderItem(Base):
     __tablename__ = "order_items"
@@ -54,7 +50,8 @@ class OrderItem(Base):
     order_id = Column(Integer, ForeignKey("orders.id"), nullable=False)
     product_id = Column(Integer, nullable=False) 
     
-    quantity = Column(Integer, nullable=False)
+    # CHANGED TO FLOAT to support decimal measurements (e.g., 1.5 KG)
+    quantity = Column(Float, nullable=False)
     price_at_purchase = Column(Float, nullable=False)
 
     order = relationship("Order", back_populates="items")
